Refactor: Outsource selection algorithm to helper
