+++ src/command/show.rs
@@ -23,80 +23,8 @@ pub struct Show {
 
 impl crate::command::Command for Show {
     fn execute(self, workdir: &Path, config: &Configuration) -> Result<(), Error> {
-        let walk_dir = |path| {
-            walkdir::WalkDir::new(path)
-                .follow_links(false)
-                .max_open(100)
-                .same_file_system(true)
-                .into_iter()
-        };
-
-        let result_dir_entry_to_pathbuf = |rde: Result<walkdir::DirEntry, _>| match rde {
-            Ok(de) => de.path().is_file().then(|| de.path().to_path_buf()).map(Ok),
-            Err(e) => Some(Err(Error::from(e))),
-        };
-
-        let is_gitkeep = |rpath: &Result<PathBuf, _>| match rpath {
-            Ok(path) => path.ends_with(".gitkeep"),
-            Err(_) => true,
-        };
-
-        let pathes = match self.selector {
-            None | Some(Selector::Unreleased) => {
-                log::debug!("Showing unreleased");
-                let unreleased_dir_path = workdir
-                    .join(config.fragment_dir())
-                    .join(crate::consts::UNRELEASED_DIR_NAME);
-                walk_dir(unreleased_dir_path)
-                    .filter_map(result_dir_entry_to_pathbuf)
-                    .filter(|r| !is_gitkeep(r))
-                    .collect::<Result<Vec<PathBuf>, Error>>()?
-            }
-            Some(Selector::Exact { exact }) => {
-                log::debug!("Showing exact {exact}");
-                let path = workdir.join(config.fragment_dir()).join(&exact);
-                if !path.exists() {
-                    return Err(Error::ExactVersionDoesNotExist { version: exact });
-                }
-                walk_dir(path)
-                    .filter_map(result_dir_entry_to_pathbuf)
-                    .filter(|r| !is_gitkeep(r))
-                    .collect::<Result<Vec<PathBuf>, Error>>()?
-            }
-            Some(Selector::Range { from, until }) => {
-                log::debug!("Showing range from {from} until {until}");
-                let from = semver::Version::parse(&from)?;
-                let until = semver::Version::parse(&until)?;
-
-                let fragment_dir_path = workdir.join(config.fragment_dir());
-                walk_dir(fragment_dir_path)
-                    .filter_entry(|de| {
-                        log::debug!("Looking at {de:?}");
-                        if de.path().is_dir() {
-                            true
-                        } else if de.path().is_file() {
-                            de.path().components().any(|comp| match comp {
-                                std::path::Component::Normal(osstr) => osstr
-                                    .to_str()
-                                    .map(|s| {
-                                        if let Ok(version) = semver::Version::parse(s) {
-                                            version > from && version < until
-                                        } else {
-                                            false
-                                        }
-                                    })
-                                    .unwrap_or(false),
-                                _ => false,
-                            })
-                        } else {
-                            false
-                        }
-                    })
-                    .filter_map(result_dir_entry_to_pathbuf)
-                    .filter(|r| !is_gitkeep(r))
-                    .collect::<Result<Vec<PathBuf>, Error>>()?
-            }
-        };
+        let pathes =
+            crate::selector::SelectorExecutor::new(self.selector.as_ref()).run(workdir, config)?;
 
         log::trace!("Looking at: {pathes:?}");
         let fragments = pathes.into_iter().map(|path| {
